# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RPA Central Monitoring API",
    "version": "3.2.0",
    "description": "Authenticated PostgreSQL-backed API for robot inventory, machine heartbeat monitoring, latest runs, and Power Automate logging. Dashboard users are provisioned directly in the database; public registration is not available."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "security": [{ "cookieAuth": [] }],
...
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Sign in with a database-provisioned account",
        "operationId": "login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in and HttpOnly session cookie set",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Delete the current session",
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "Signed out",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OkResponse" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get the signed-in user",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "Current session user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/api/dashboard": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Get robots, active machines, and latest runs",
        "description": "Returns all rpa_robot rows so turned-off robots remain visible, active rpa_machine rows, the latest rpa_robot_run for each robot, and events for those runs.",
        "operationId": "getDashboard",
        "responses": {
          "200": {
            "description": "Dashboard data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DashboardResponse" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/history": {
      "get": {
        "tags": ["History"],
        "summary": "Search robot run history",
        "description": "Returns newest-first run records joined with robot display information.",
        "operationId": "getRunHistory",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Robot name/code, run ID, machine, input reference, or error text",
            "schema": { "type": "string", "maxLength": 200 }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "oneOf": [
                { "$ref": "#/components/schemas/RunStatus" },
                { "type": "string", "enum": ["ALL"] }
              ],
              "default": "ALL"
            }
          },
          {
            "name": "startedFrom",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "startedTo",
            "in": "query",
            "schema": { "type": "string", "format": "date-time" }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "minimum": 1, "maximum": 500, "default": 200 }
          }
        ],
        "responses": {
          "200": {
            "description": "Filtered run history",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RunHistoryResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/robots": {
      "post": {
        "tags": ["Robots"],
        "summary": "Create a robot with an automatic code",
        "description": "Creates a robot and assigns the next unique RPA-000001 style code on the server. Account Name and AnyDesk ID are display metadata only; no passwords are accepted.",
        "operationId": "createRobot",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RobotCreateRequest" },
              "example": {
                "robotName": "ITZONE Receipt Bot",
                "robotType": "CLOUD_DESKTOP",
                "powerAutomateEnvironmentId": "Default-xxxxxxxx",
                "cloudFlowId": "xxxxxxxx",
                "cloudFlowName": "ITZONE Receipt Main Flow",
                "cloudFlowUrl": "https://make.powerautomate.com/manage/environments/Default-xxxxxxxx/flows/xxxxxxxx/details",
                "desktopFlowId": "xxxxxxxx",
                "desktopFlowName": "ITZONE Receipt PAD",
                "desktopFlowUrl": "https://make.powerautomate.com/manage/environments/Default-xxxxxxxx/uiflows/xxxxxxxx/details",
                "accountName": "Robot Account 02",
                "machineName": "BOT-PC-02",
                "machineIp": "10.0.0.22",
                "anydeskId": "123 456 789"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created or updated robot",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RobotResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/robots/{robotId}/active": {
      "patch": {
        "tags": ["Robots"],
        "summary": "Turn a robot on or off",
        "description": "ADMIN only. Turned-off robots remain visible, but new logger START requests return 409 Conflict. Existing RUNNING records may still be completed.",
        "operationId": "updateRobotActive",
        "parameters": [
          {
            "name": "robotId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RobotActiveUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Robot availability updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RobotResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/robots/{robotId}/power-automate": {
      "patch": {
        "tags": ["Robots"],
        "summary": "Update a robot's Power Automate links",
        "description": "ADMIN only. Stores separate Cloud Flow and Desktop Flow targets plus the account name shown to dashboard users.",
        "operationId": "updateRobotPowerAutomate",
        "parameters": [
          {
            "name": "robotId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RobotPowerAutomateUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Power Automate settings updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RobotResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/machines": {
      "get": {
        "tags": ["Machines"],
        "summary": "List machines with calculated availability",
        "description": "Status reports machine availability only, derived purely from the heartbeat. NOT_CONNECTED means the heartbeat agent has never reported. OFFLINE means the last heartbeat is older than MACHINE_OFFLINE_SECONDS. ONLINE means a heartbeat arrived inside that window. Robot workload is reported separately through runningRunCount and staleRunningCount.",
        "operationId": "getMachines",
        "responses": {
          "200": {
            "description": "Machine status list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MachineListResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/machines/heartbeat": {
      "post": {
        "tags": ["Machines"],
        "summary": "Record a machine heartbeat",
        "description": "Creates the machine when needed and refreshes its IP, AnyDesk ID, metadata, and server-side heartbeat timestamp.",
        "operationId": "recordMachineHeartbeat",
        "security": [{ "cookieAuth": [] }, { "apiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MachineHeartbeatRequest" },
              "example": {
                "machineName": "BOT-PC-02",
                "machineIp": "10.0.0.22",
                "anydeskId": "123 456 789",
                "metadata": {
                  "os": "Microsoft Windows NT 10.0.26100.0",
                  "agent": "powershell-heartbeat"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Heartbeat recorded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MachineResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/logger/start": {
      "post": {
        "tags": ["Logging"],
        "summary": "Create a RUNNING robot run",
        "operationId": "startRobotRun",
        "security": [{ "cookieAuth": [] }, { "apiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoggerStartRequest" },
              "example": {
                "robotCode": "RPA-000001",
                "cloudFlowRunId": "flow-run-001",
                "machineName": "BOT-PC-02"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "RUNNING record created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RobotRunResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rpa_session",
        "description": "HttpOnly session cookie set by /api/auth/login. The browser sends it automatically."
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RPA-API-Key",
        "description": "Optional deployment secret for Power Automate logging endpoints."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } }
        }
      },
...
    },
    "schemas": {
...
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "maxLength": 100 },
          "password": { "type": "string", "format": "password", "minLength": 12, "maxLength": 256, "writeOnly": true }
        }
      },
      "AuthResponse": {
        "type": "object",
        "required": ["user", "expiresAt"],
        "properties": {
          "user": { "$ref": "#/components/schemas/User" },
          "expiresAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "OkResponse": {
        "type": "object",
        "required": ["ok"],
        "properties": {
          "ok": { "type": "boolean", "example": true }
        }
      },
...
    }
  }
}